make averages.py file
