+++ pybamm/__init__.py
@@ -80,11 +80,13 @@ from .expression_tree.concatenations import *
 from .expression_tree.array import Array, linspace, meshgrid
 from .expression_tree.matrix import Matrix
 from .expression_tree.unary_operators import *
+from .expression_tree.averages import *
+from .expression_tree.averages import _BaseAverage
+from .expression_tree.broadcasts import *
 from .expression_tree.functions import *
 from .expression_tree.interpolant import Interpolant
 from .expression_tree.input_parameter import InputParameter
 from .expression_tree.parameter import Parameter, FunctionParameter
-from .expression_tree.broadcasts import *
 from .expression_tree.scalar import Scalar
 from .expression_tree.variable import *
 from .expression_tree.independent_variable import *
